// src/modules/invoice/usecase/generate/generate-invoice.dto.ts
export interface GenerateInvoiceUseCaseInputDTO {
    name: string;
    document: string;
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipcode: string;
    items: {
        id: string;
        name: string;
        price: number;
    }[];
}

export interface GenerateInvoiceUseCaseOutputDTO {
    id: string;
    name: string;
    document: string;
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipcode: string;
    items: {
        id: string;
        name: string;
        price: number;
    }[];
    total: number;
}